some initial clean-up for animation and animator, animation_editor still needs some much needed love
